ShootApp v1 ready for release

+++ frontend/src/views/CompetitionsView.vue
@@ -11,6 +11,8 @@ const competitions = ref([])
 const showForm = ref(false)
 const editMode = ref(false)
 const visibleIds = ref({})
+const errorMessage = ref('')
+const successMessage = ref('')
 
 const newCompetition = ref({
   name: '',
@@ -33,6 +35,7 @@ async function handleSaveCompetition() {
       await updateCompetition(newCompetition.value)
     } else {
       await createCompetition(newCompetition.value)
+      successMessage.value = 'Súťaž bola vytvorená úspešne !'
     }
 
     competitions.value = await getCompetitions()
@@ -47,15 +50,30 @@ async function handleSaveCompetition() {
     showForm.value = false
 
   } catch (error) {
-    if(error.response?.data?.message) {
-      alert(error.response.data.message)
+
+      if (error.response?.data?.errors) {
+        errorMessage.value = error.response.data.errors.join(', ')
+        return
+      }
+
+      if (error.response?.data?.message) {
+        errorMessage.value = error.response.data.message
+        return
+      }
+
+      errorMessage.value = 'Unexpected error'
     }
-  }
 }
 
 async function handleDeleteCompetition(id) {
   try {
+
+    if (!confirm('Ste si istý že chcete vymazať túto súťaž ?')) {
+      return
+    }
+
     await deleteCompetition(id)
+
     competitions.value = await getCompetitions()
   } catch (error) {
       console.error(error)
@@ -96,6 +114,20 @@ async function handleDeleteCompetition(id) {
       </button>
     </div>
 
+    <div
+      v-if="errorMessage"
+      class="error-box"
+    >
+      {{ errorMessage }}
+    </div>
+
+     <div
+          v-if="successMessage"
+          class="success-box"
+        >
+          {{ successMessage }}
+     </div>
+
     <div v-if="showForm" class="form-container">
       <input v-model="newCompetition.name" placeholder="Názov" />
       <input v-model="newCompetition.place" placeholder="Miesto" />
@@ -106,6 +138,8 @@ async function handleDeleteCompetition(id) {
       </button>
 
     </div>
+
+
 
     <div class="table-container">
       <table>
@@ -198,6 +232,11 @@ table {
   border-collapse: collapse;
 }
 
+tbody tr:hover {
+  background: #f9fafb;
+}
+
+
 th {
   background: #f3f4f6;
   text-align: left;
@@ -265,6 +304,24 @@ td {
   margin-left: 10px;
   margin-right: 10px;
   font-weight: bold;
+}
+
+.success-box {
+  margin-bottom: 20px;
+  padding: 12px;
+  border-radius: 8px;
+  background: #dcfce7;
+  color: #166534;
+  border: 1px solid #bbf7d0;
+}
+
+.error-box {
+  margin-bottom: 20px;
+  padding: 12px;
+  border-radius: 8px;
+  background: #fee2e2;
+  color: #991b1b;
+  border: 1px solid #fecaca;
 }
 
 </style>
