+++ frontend/src/views/RoundsView.vue
@@ -13,6 +13,10 @@ const showForm = ref(false)
 
 const visibleIds = ref({})
 
+const errorMessage = ref('')
+const successMessage = ref('')
+
+
 const newRound = ref({
   compId: '',
   name: '',
@@ -55,17 +59,26 @@ async function handleSaveRound() {
 
   } catch (error) {
 
-    if (error.response?.data?.message) {
-      alert(error.response.data.message)
-    } else {
-      alert('Unexpected error')
-    }
+              if (error.response?.data?.errors) {
+                errorMessage.value = error.response.data.errors.join(', ')
+                return
+              }
 
-  }
+              if (error.response?.data?.message) {
+                errorMessage.value = error.response.data.message
+                return
+              }
+
+              errorMessage.value = 'Unexpected error'
+            }
 }
 
 async function handleDeleteRound(id) {
   try {
+
+    if (!confirm('Ste si istý že chcete vymazať toto kolo ?')) {
+      return
+    }
 
     await deleteRound(id)
 
@@ -103,6 +116,21 @@ function toggleId(id) {
         + Pridať kolo
       </button>
     </div>
+
+<div
+          v-if="errorMessage"
+          class="error-box"
+        >
+          {{ errorMessage }}
+        </div>
+
+         <div
+              v-if="successMessage"
+              class="success-box"
+            >
+              {{ successMessage }}
+         </div>
+
 
     <div
       v-if="showForm"
@@ -248,6 +276,10 @@ table {
   border-collapse: collapse;
 }
 
+tbody tr:hover {
+  background: #f9fafb;
+}
+
 th {
   background: #f3f4f6;
   text-align: left;
@@ -282,5 +314,23 @@ td {
   margin-left: 10px;
   margin-right: 10px;
   font-weight: bold;
+}
+
+.success-box {
+  margin-bottom: 20px;
+  padding: 12px;
+  border-radius: 8px;
+  background: #dcfce7;
+  color: #166534;
+  border: 1px solid #bbf7d0;
+}
+
+.error-box {
+  margin-bottom: 20px;
+  padding: 12px;
+  border-radius: 8px;
+  background: #fee2e2;
+  color: #991b1b;
+  border: 1px solid #fecaca;
 }
 </style>
