+++ frontend/src/views/UsersView.vue
@@ -11,6 +11,8 @@ const users = ref([])
 const showForm = ref(false)
 const editMode = ref(false)
 const visibleIds = ref({})
+const errorMessage = ref('')
+const successMessage = ref('')
 
 const newUser = ref({
   name: '',
@@ -34,6 +36,7 @@ async function handleSaveUser() {
       await updateUser(newUser.value)
     } else {
       await createUser(newUser.value)
+      successMessage.value = 'Používateľ bol vytvorený úspešne !'
     }
 
     users.value = await getUsers()
@@ -49,14 +52,26 @@ async function handleSaveUser() {
     showForm.value = false
 
   } catch (error) {
-    if(error.response?.data?.message) {
-      alert(error.response.data.message)
+
+      if (error.response?.data?.errors) {
+        errorMessage.value = error.response.data.errors.join(', ')
+        return
+      }
+
+      if (error.response?.data?.message) {
+        errorMessage.value = error.response.data.message
+        return
+      }
+
+      errorMessage.value = 'Unexpected error'
     }
-  }
 }
 
 async function handleDeleteUser(id) {
   try {
+    if (!confirm('Ste si istý že chcete vymazať tohto používatela ?')) {
+      return
+    }
     await deleteUser(id)
     users.value = await getUsers()
   } catch (error) {
@@ -97,6 +112,20 @@ async function handleDeleteUser(id) {
       <button class="create-btn" @click="showForm = !showForm">
         + Pridať používateľa
       </button>
+    </div>
+
+    <div
+      v-if="errorMessage"
+      class="error-box"
+    >
+      {{ errorMessage }}
+    </div>
+
+    <div
+      v-if="successMessage"
+      class="success-box"
+    >
+      {{ successMessage }}
     </div>
 
     <div v-if="showForm" class="form-container">
@@ -205,6 +234,11 @@ table {
   border-collapse: collapse;
 }
 
+tbody tr:hover {
+  background: #f9fafb;
+}
+
+
 th {
   background: #f3f4f6;
   text-align: left;
@@ -272,6 +306,24 @@ td {
   margin-left: 10px;
   margin-right: 10px;
   font-weight: bold;
+}
+
+.success-box {
+  margin-bottom: 20px;
+  padding: 12px;
+  border-radius: 8px;
+  background: #dcfce7;
+  color: #166534;
+  border: 1px solid #bbf7d0;
+}
+
+.error-box {
+  margin-bottom: 20px;
+  padding: 12px;
+  border-radius: 8px;
+  background: #fee2e2;
+  color: #991b1b;
+  border: 1px solid #fecaca;
 }
 
 </style>
